some test utility consolidation but a new issue should be created
