+++ examples/integration-scripts/singlesig-ixn.test.ts
@@ -1,10 +1,11 @@
 import { EventResult, SignifyClient } from 'signify-ts';
 import {
+    assertOperations,
     getOrCreateClients,
     getOrCreateContact,
     getOrCreateIdentifier,
-} from './utils/test-setup';
-import { assertOperations, waitOperation } from './utils/test-util';
+    waitOperation,
+} from './utils/test-util';
 
 let client1: SignifyClient, client2: SignifyClient;
 let name1_id: string, name1_oobi: string;
